Add the way to debug calibrated high side data.
When connecting to the upper computer and receiving a command to write flash,
function is25pl032_flash_get_calibrated_high_data in the ie_task. c file will
cause flash write failure.

So this is just for the convenience of debugging high edge changes, and users
need to comment out is25pl032_flash_get_calibrated_high_data in the ie_task. c
file when using it.