TG: Optimize vibration detection function
Issue: dmc#5

Add vibration dectation task. Add median filtering and mean filtering.
The vibration data of the ADXL357 is collected every 100 milliseconds,
and the delay of the vibration trigger setting is added, that is, if
the average value of the vibration data after median filtering and mean
filtering is greater than the threshold within 8 seconds, it indicates
vibration.